Eng 1310 streaming clean up (#6540)
