+++ tests/ops/service/privacy_request/dsr_report_builder/test_dsr_report_builder.py
@@ -1,5 +1,4 @@
 import io
-import re
 import threading
 import time
 import zipfile
@@ -18,465 +17,7 @@ from fides.api.service.privacy_request.dsr_package.dsr_report_builder import (
     DSRReportBuilder,
 )
 
-
-class TestDSRReportBuilderBase:
-    """Base class with common test utilities"""
-
-    @staticmethod
-    def assert_file_in_zip(zip_file, file_path, content=None, is_binary=False):
-        """Assert that a file exists in the zip file and optionally check its content."""
-        assert file_path in zip_file.namelist()
-        if content is not None:
-            if is_binary:
-                assert zip_file.read(file_path) == content
-            else:
-                assert zip_file.read(file_path).decode("utf-8") == content
-
-    @staticmethod
-    def assert_html_contains(html_content, *strings):
-        """Assert that the HTML content contains all the given strings."""
-        for string in strings:
-            assert string in html_content
-
-    @staticmethod
-    def assert_html_not_contains(html_content, *strings):
-        """Assert that the HTML content does not contain any of the given strings."""
-        for string in strings:
-            assert string not in html_content
-
-    @staticmethod
-    def extract_table_values(html_content: str) -> dict[str, str]:
-        """Extract values from HTML table cells."""
-        values = {}
-        # Find all table rows (div with class table-row), being lenient with whitespace
-        rows = re.findall(
-            r'<div\s+class="table-row">\s*<div\s+class="table-cell">(.*?)</div>\s*<div\s+class="table-cell">(.*?)</div>\s*</div>',
-            html_content,
-            re.DOTALL,
-        )
-
-        for i, (key_cell, value_cell) in enumerate(rows):
-            # Extract the key
-            key = re.sub(r"<[^>]+>", "", key_cell).strip()
-
-            # Extract the value from pre tag
-            value_match = re.search(r"<pre>(.*?)</pre>", value_cell, re.DOTALL)
-            if value_match:
-                value = value_match.group(1).strip()
-                # Unescape HTML entities
-                value = value.replace("&#34;", '"')
-                if key and value:  # Only add non-empty values
-                    values[key] = value
-
-        return values
-
-
-@pytest.fixture
-def common_webhook_data():
-    """Common webhook data fixture with all possible fields"""
-    return {
-        "email": "test@example.com",
-        "name": "Test User",
-        "last_name": "Test User",
-        "phone": "+1234567890",
-        "system_name": "test_system",
-    }
-
-
-@pytest.fixture
-def common_attachment_config():
-    """Common attachment configuration fixture with all attachment types"""
-    return {
-        "text": {
-            "file_name": "test.txt",
-            "file_size": 1024,
-            "content_type": "text/plain",
-            "fileobj": BytesIO(b"test content"),
-            "download_url": "http://example.com/test.txt",
-        },
-        "binary": {
-            "file_name": "test.pdf",
-            "file_size": 2048,
-            "content_type": "application/pdf",
-            "fileobj": BytesIO(b"test content"),
-            "download_url": "http://example.com/test.pdf",
-        },
-        "image": {
-            "file_name": "test.jpg",
-            "file_size": 3072,
-            "content_type": "image/jpeg",
-            "fileobj": BytesIO(b"fake image content"),
-            "download_url": "http://example.com/test.jpg",
-        },
-    }
-
-
-@pytest.fixture
-def webhook_variants(common_webhook_data, common_attachment_config):
-    """Fixture providing different webhook data variants"""
-    return {
-        "basic": common_webhook_data,
-        "empty_attachments": {**common_webhook_data, "attachments": []},
-        "with_attachments": {
-            **common_webhook_data,
-            "attachments": [
-                common_attachment_config["text"],
-                common_attachment_config["binary"],
-            ],
-        },
-        "malformed": {
-            **common_webhook_data,
-            "attachments": "not a list",
-            "nested_data": {"attachments": None},
-        },
-    }
-
-
-@pytest.fixture
-def common_file_assertions():
-    """Common file existence assertions fixture"""
-    return {
-        "required": [
-            "welcome.html",
-            "data/main.css",
-            "data/back.svg",
-        ],
-        "attachment_extensions": [".txt", ".pdf", ".jpg", ".png"],
-        "index_files": [
-            "index.html",
-            "attachments/index.html",
-        ],
-    }
-
-
-@pytest.fixture
-def common_assertions(common_file_assertions):
-    """Common assertions for DSR report builder tests"""
-    return {
-        "files": common_file_assertions,
-        "paths": {
-            "webhook_dir": "data/test_system/test_webhook",
-            "webhook_dir2": "data/test_system2/test_webhook2",
-            "attachments_dir": "attachments",
-            "manual_webhook_dir": "data/manual/test_webhook",
-            "welcome_path": "welcome.html",
-            "attachments_index": "attachments/index.html",
-            "back_svg_path": "data/back.svg",
-            "collection_dir": "data/dataset/collection",
-            "collection_index": "data/dataset/collection/index.html",
-            "item_index": "data/dataset/collection/item_index.html",
-        },
-        "html": {
-            "attachment_link": "attachment-link",
-            "attachment_size": "attachment-size",
-            "attachment_name": "attachment-name",
-        },
-    }
-
-
-class TestDSRReportBuilderAttachments(TestDSRReportBuilderBase):
-    """Tests for attachment handling in DSR reports"""
-
-    def test_webhook_attachments(
-        self,
-        privacy_request: PrivacyRequest,
-        webhook_variants,
-        common_assertions,
-        common_attachment_config,
-    ):
-        """Test handling of attachments in webhook data"""
-        dsr_data = {
-            "test_webhook": [webhook_variants["basic"]],
-            "attachments": [
-                common_attachment_config["text"],
-                common_attachment_config["binary"],
-            ],
-        }
-
-        builder = DSRReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
-        report = builder.generate()
-
-        with zipfile.ZipFile(io.BytesIO(report.getvalue())) as zip_file:
-            # Verify that the attachments index file exists
-            self.assert_file_in_zip(
-                zip_file, f"{common_assertions['paths']['attachments_dir']}/index.html"
-            )
-
-            # Read and verify the content of the attachments index file
-            attachments_content = zip_file.read(
-                f"{common_assertions['paths']['attachments_dir']}/index.html"
-            ).decode("utf-8")
-
-            # Verify both attachment links are present with their details
-            self.assert_html_contains(
-                attachments_content,
-                common_attachment_config["text"]["file_name"],
-                common_attachment_config["text"]["download_url"],
-                "1.0 KB",  # Assuming the file size is formatted as 1.0 KB
-            )
-            self.assert_html_contains(
-                attachments_content,
-                common_attachment_config["binary"]["file_name"],
-                common_attachment_config["binary"]["download_url"],
-                "2.0 KB",  # Assuming the file size is formatted as 2.0 KB
-            )
-
-    def test_streaming_attachment_links(
-        self,
-        privacy_request: PrivacyRequest,
-        common_attachment_config,
-        common_assertions,
-    ):
-        """Test that when streaming is enabled, attachment links point to local attachments directory"""
-        dsr_data = {
-            "attachments": [
-                common_attachment_config["text"],
-                common_attachment_config["binary"],
-            ],
-        }
-
-        # Test with streaming enabled
-        builder_streaming = DSRReportBuilder(
-            privacy_request=privacy_request, dsr_data=dsr_data, enable_streaming=True
-        )
-        report_streaming = builder_streaming.generate()
-
-        with zipfile.ZipFile(io.BytesIO(report_streaming.getvalue())) as zip_file:
-            # Read the attachments index file
-            attachments_content = zip_file.read(
-                f"{common_assertions['paths']['attachments_dir']}/index.html"
-            ).decode("utf-8")
-
-            # Verify that links point to local attachments directory
-            self.assert_html_contains(
-                attachments_content,
-                common_attachment_config["text"]["file_name"],
-                common_attachment_config["text"][
-                    "file_name"
-                ],  # Just filename, not full path
-                "1.0 KB",
-            )
-            self.assert_html_contains(
-                attachments_content,
-                common_attachment_config["binary"]["file_name"],
-                common_attachment_config["binary"][
-                    "file_name"
-                ],  # Just filename, not full path
-                "2.0 KB",
-            )
-
-            # Verify that original download URLs are NOT present
-            self.assert_html_not_contains(
-                attachments_content,
-                common_attachment_config["text"]["download_url"],
-                common_attachment_config["binary"]["download_url"],
-            )
-
-        # Test with streaming disabled (default behavior)
-        builder_default = DSRReportBuilder(
-            privacy_request=privacy_request, dsr_data=dsr_data, enable_streaming=False
-        )
-        report_default = builder_default.generate()
-
-        with zipfile.ZipFile(io.BytesIO(report_default.getvalue())) as zip_file:
-            # Read the attachments index file
-            attachments_content = zip_file.read(
-                f"{common_assertions['paths']['attachments_dir']}/index.html"
-            ).decode("utf-8")
-
-            # Verify that links point to original download URLs
-            self.assert_html_contains(
-                attachments_content,
-                common_attachment_config["text"]["file_name"],
-                common_attachment_config["text"]["download_url"],
-                "1.0 KB",
-            )
-            self.assert_html_contains(
-                attachments_content,
-                common_attachment_config["binary"]["file_name"],
-                common_attachment_config["binary"]["download_url"],
-                "2.0 KB",
-            )
-
-    def test_ttl_display_in_templates(
-        self,
-        privacy_request: PrivacyRequest,
-        common_attachment_config,
-        common_assertions,
-    ):
-        """Test that TTL is displayed correctly in templates using the environment variable"""
-        dsr_data = {
-            "attachments": [
-                common_attachment_config["text"],
-            ],
-        }
-
-        builder = DSRReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
-        report = builder.generate()
-
-        with zipfile.ZipFile(io.BytesIO(report.getvalue())) as zip_file:
-            # Read the attachments index file
-            attachments_content = zip_file.read(
-                f"{common_assertions['paths']['attachments_dir']}/index.html"
-            ).decode("utf-8")
-
-            # Verify that the TTL is displayed (should be 5 days by default)
-            # The exact number will depend on the CONFIG.security.subject_request_download_link_ttl_seconds value
-            self.assert_html_contains(
-                attachments_content, "download links will expire in"
-            )
-            self.assert_html_contains(attachments_content, "days")
-
-            # Verify that the hardcoded "7 days" is not present
-            self.assert_html_not_contains(attachments_content, "7 days")
-
-    def test_ttl_display_in_collection_templates(
-        self,
-        privacy_request: PrivacyRequest,
-        common_assertions,
-    ):
-        """Test that TTL is displayed correctly in collection templates using the environment variable"""
-        dsr_data = {
-            "test_dataset:test_collection": [
-                {
-                    "id": 1,
-                    "name": "Test Item",
-                    "attachments": {
-                        "test_file.txt": {
-                            "url": "https://example.com/test_file.txt",
-                            "size": "1.0 KB",
-                        }
-                    },
-                }
-            ]
-        }
-
-        builder = DSRReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
-        report = builder.generate()
-
-        with zipfile.ZipFile(io.BytesIO(report.getvalue())) as zip_file:
-            # Read the collection index file
-            collection_content = zip_file.read(
-                "data/test_dataset/test_collection/index.html"
-            ).decode("utf-8")
-
-            # Verify that the TTL is displayed
-            self.assert_html_contains(
-                collection_content, "download links will expire in"
-            )
-            self.assert_html_contains(collection_content, "days")
-
-            # Verify that the hardcoded "7 days" is not present
-            self.assert_html_not_contains(collection_content, "7 days")
-
-    def test_multiple_webhook_attachments(
-        self, privacy_request: PrivacyRequest, webhook_variants, common_assertions
-    ):
-        """Test handling of attachments across multiple webhooks"""
-        dsr_data = {
-            "test_webhook": [
-                {
-                    **webhook_variants["basic"],
-                    "attachments": [
-                        webhook_variants["with_attachments"]["attachments"][0]
-                    ],
-                }
-            ],
-            "test_webhook2": [
-                {
-                    "phone": webhook_variants["basic"]["phone"],
-                    "attachments": [
-                        webhook_variants["with_attachments"]["attachments"][1]
-                    ],
-                    "system_name": "test_system2",
-                }
-            ],
-            "attachments": [webhook_variants["with_attachments"]["attachments"][0]],
-        }
-
-        builder = DSRReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
-        report = builder.generate()
-
-        with zipfile.ZipFile(io.BytesIO(report.getvalue())) as zip_file:
-            # Verify that webhook index files exist
-            self.assert_file_in_zip(
-                zip_file, f"{common_assertions['paths']['webhook_dir']}/index.html"
-            )
-            self.assert_file_in_zip(
-                zip_file, f"{common_assertions['paths']['webhook_dir2']}/index.html"
-            )
-
-            # Read and verify the content of the webhook index files
-            webhook1_content = zip_file.read(
-                f"{common_assertions['paths']['webhook_dir']}/index.html"
-            ).decode("utf-8")
-            webhook2_content = zip_file.read(
-                f"{common_assertions['paths']['webhook_dir2']}/index.html"
-            ).decode("utf-8")
-
-            # Verify attachment links in webhook1
-            self.assert_html_contains(
-                webhook1_content,
-                webhook_variants["with_attachments"]["attachments"][0]["file_name"],
-                webhook_variants["with_attachments"]["attachments"][0]["download_url"],
-                "1.0 KB",  # Assuming the file size is formatted as 1.0 KB
-            )
-
-            # Verify attachment links in webhook2
-            self.assert_html_contains(
-                webhook2_content,
-                webhook_variants["with_attachments"]["attachments"][1]["file_name"],
-                webhook_variants["with_attachments"]["attachments"][1]["download_url"],
-                "2.0 KB",  # Assuming the file size is formatted as 2.0 KB
-            )
-
-    @pytest.mark.parametrize("webhook_type", ["empty_attachments", "malformed"])
-    def test_webhook_variants(
-        self,
-        privacy_request: PrivacyRequest,
-        webhook_variants,
-        common_assertions,
-        webhook_type,
-    ):
-        """Test different webhook data variants"""
-        dsr_data = {
-            "manual:test_webhook": [webhook_variants[webhook_type]],
-        }
-
-        builder = DSRReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
-        report = builder.generate()
-
-        with zipfile.ZipFile(io.BytesIO(report.getvalue())) as zip_file:
-            # Verify that the report was still generated
-            self.assert_file_in_zip(
-                zip_file,
-                f"{common_assertions['paths']['manual_webhook_dir']}/index.html",
-            )
-            manual_data = zip_file.read(
-                f"{common_assertions['paths']['manual_webhook_dir']}/index.html"
-            ).decode("utf-8")
-            self.assert_html_contains(
-                manual_data, webhook_variants[webhook_type]["email"]
-            )
-
-            # Verify that attachments directory exists but contains no actual attachment files
-            # (only the index.html file should be present)
-            attachment_files = [
-                name
-                for name in zip_file.namelist()
-                if name.startswith(f"{common_assertions['paths']['attachments_dir']}/")
-            ]
-            # Should only have the index.html file, no actual attachment files
-            assert len(attachment_files) == 1
-            assert (
-                attachment_files[0]
-                == f"{common_assertions['paths']['attachments_dir']}/index.html"
-            )
-
-            # Verify that the attachments index page exists and is accessible
-            self.assert_file_in_zip(
-                zip_file, f"{common_assertions['paths']['attachments_dir']}/index.html"
-            )
+from .conftest import TestDSRReportBuilderBase
 
 
 class TestDSRReportBuilderDataStructure:
@@ -908,193 +449,6 @@ class TestDSRReportBuilder(TestDSRReportBuilderBase):
             )
 
 
-class TestDSRReportBuilderAttachmentHandling:
-    """Tests for DSR report builder's attachment handling functions"""
-
-    def test_handle_attachment_text(self, privacy_request: PrivacyRequest):
-        """Test handling of text attachments with download URLs"""
-        builder = DSRReportBuilder(privacy_request, {})
-        attachments = [
-            {
-                "file_name": "test.txt",
-                "download_url": "https://example.com/test.txt",
-                "file_size": 1024,
-            }
-        ]
-
-        result = builder._write_attachment_content(attachments, "attachments")
-
-        assert "test.txt" in result
-        assert result["test.txt"]["url"] == "https://example.com/test.txt"
-        assert result["test.txt"]["size"] == "1.0 KB"
-
-    def test_handle_attachment_binary(self, privacy_request: PrivacyRequest):
-        """Test handling of binary attachments with download URLs"""
-        builder = DSRReportBuilder(privacy_request, {})
-        attachments = [
-            {
-                "file_name": "test.pdf",
-                "download_url": "https://example.com/test.pdf",
-                "file_size": 2048576,  # 2MB
-            }
-        ]
-
-        result = builder._write_attachment_content(attachments, "attachments")
-
-        assert "test.pdf" in result
-        assert result["test.pdf"]["url"] == "https://example.com/test.pdf"
-        assert result["test.pdf"]["size"] == "2.0 MB"
-
-    def test_handle_attachment_no_content(self, privacy_request: PrivacyRequest):
-        """Test handling of attachments with missing content"""
-        builder = DSRReportBuilder(privacy_request, {})
-        attachments = [
-            {"file_name": "test.txt", "download_url": None, "file_size": None}
-        ]
-
-        result = builder._write_attachment_content(attachments, "attachments")
-        assert not result  # Should return empty dict for invalid attachments
-
-    def test_add_attachments_top_level(self, privacy_request: PrivacyRequest):
-        """Test adding top-level attachments to the report"""
-        attachments = [
-            {
-                "file_name": "test1.txt",
-                "download_url": "https://example.com/test1.txt",
-                "file_size": 1024,
-            },
-            {
-                "file_name": "test2.pdf",
-                "download_url": "https://example.com/test2.pdf",
-                "file_size": 2048576,
-            },
-        ]
-
-        dsr_data = {"attachments": attachments}
-        builder = DSRReportBuilder(privacy_request, dsr_data)
-        zip_file = builder.generate()
-
-        # Create a ZipFile object from the BytesIO
-        with zipfile.ZipFile(zip_file) as zip_file_obj:
-            # Check that attachments index was created
-            TestDSRReportBuilderBase.assert_file_in_zip(
-                zip_file_obj, "attachments/index.html"
-            )
-
-            # Check that the HTML contains the attachment links
-            html_content = zip_file_obj.read("attachments/index.html").decode("utf-8")
-            TestDSRReportBuilderBase.assert_html_contains(
-                html_content,
-                "test1.txt",
-                "test2.pdf",
-                "https://example.com/test1.txt",
-                "https://example.com/test2.pdf",
-                "1.0 KB",
-                "2.0 MB",
-            )
-
-    def test_process_item_attachments(self, privacy_request: PrivacyRequest):
-        """Test processing attachments within collection items"""
-        dsr_data = {
-            "dataset1:collection1": [
-                {
-                    "id": 1,
-                    "name": "Item 1",
-                    "attachments": [
-                        {
-                            "file_name": "item1.txt",
-                            "download_url": "https://example.com/item1.txt",
-                            "file_size": 1024,
-                        }
-                    ],
-                }
-            ]
-        }
-
-        builder = DSRReportBuilder(privacy_request, dsr_data)
-        zip_file = builder.generate()
-
-        # Create a ZipFile object from the BytesIO
-        with zipfile.ZipFile(zip_file) as zip_file_obj:
-            # Check that collection index was created
-            TestDSRReportBuilderBase.assert_file_in_zip(
-                zip_file_obj, "data/dataset1/collection1/index.html"
-            )
-
-            # Check that the HTML contains the attachment link
-            html_content = zip_file_obj.read(
-                "data/dataset1/collection1/index.html"
-            ).decode("utf-8")
-            TestDSRReportBuilderBase.assert_html_contains(
-                html_content, "item1.txt", "https://example.com/item1.txt", "1.0 KB"
-            )
-
-    def test_handle_duplicate_filenames(self, privacy_request: PrivacyRequest):
-        """Test handling of duplicate filenames across all directories"""
-        builder = DSRReportBuilder(privacy_request, {})
-        attachments = [
-            {
-                "file_name": "test.txt",
-                "download_url": "https://example.com/test1.txt",
-                "file_size": 1024,
-            },
-            {
-                "file_name": "test.txt",
-                "download_url": "https://example.com/test2.txt",
-                "file_size": 2048,
-            },
-            {
-                "file_name": "test.txt",
-                "download_url": "https://example.com/test3.txt",
-                "file_size": 3072,
-            },
-        ]
-
-        result = builder._write_attachment_content(attachments, "attachments")
-
-        # Verify that all files get unique names
-        assert "test.txt" in result  # First file keeps original name
-        assert "test_1.txt" in result  # Second file gets _1 suffix
-        assert "test_2.txt" in result  # Third file gets _2 suffix
-        assert result["test.txt"]["url"] == "https://example.com/test1.txt"
-        assert result["test_1.txt"]["url"] == "https://example.com/test2.txt"
-        assert result["test_2.txt"]["url"] == "https://example.com/test3.txt"
-
-    def test_handle_duplicate_filenames_across_directories(
-        self, privacy_request: PrivacyRequest
-    ):
-        """Test handling of duplicate filenames across different directories"""
-        builder = DSRReportBuilder(privacy_request, {})
-
-        # Test attachments in different directories
-        attachments1 = [
-            {
-                "file_name": "test.txt",
-                "download_url": "https://example.com/test1.txt",
-                "file_size": 1024,
-            }
-        ]
-        attachments2 = [
-            {
-                "file_name": "test.txt",
-                "download_url": "https://example.com/test2.txt",
-                "file_size": 2048,
-            }
-        ]
-
-        # Process attachments in sequence
-        result1 = builder._write_attachment_content(attachments1, "attachments")
-        result2 = builder._write_attachment_content(
-            attachments2, "data/dataset1/collection1"
-        )
-
-        # Verify that the second file gets a unique name
-        assert "test.txt" in result1  # First file keeps original name
-        assert "test_1.txt" in result2  # Second file gets _1 suffix
-        assert result1["test.txt"]["url"] == "https://example.com/test1.txt"
-        assert result2["test_1.txt"]["url"] == "https://example.com/test2.txt"
-
-
 class TestDSRReportBuilderDatasetHandling:
     """Tests for DSR report builder's dataset handling functions"""
 
@@ -1132,109 +486,6 @@ class TestDSRReportBuilderDatasetHandling:
             content = zip_file.read("data/dataset1/index.html").decode("utf-8")
             assert "dataset1" in content
             assert "collection" not in content
-
-
-class TestDSRReportBuilderAttachmentContentWriting:
-    """Tests for DSR report builder's attachment content writing functionality"""
-
-    @pytest.mark.parametrize(
-        "file_name,download_url,file_size,expected_size",
-        [
-            ("test.txt", "https://example.com/test.txt", 1024, "1.0 KB"),
-            ("test.pdf", "https://example.com/test.pdf", 2048576, "2.0 MB"),
-            ("test.txt", "https://example.com/test.txt", None, "Unknown"),
-        ],
-    )
-    def test_write_attachment_content_parametrized(
-        self,
-        privacy_request: PrivacyRequest,
-        file_name,
-        download_url,
-        file_size,
-        expected_size,
-    ):
-        """Test attachment content writing with various file types and sizes"""
-        builder = DSRReportBuilder(privacy_request, {})
-        attachments = [
-            {
-                "file_name": file_name,
-                "download_url": download_url,
-                "file_size": file_size,
-            }
-        ]
-
-        result = builder._write_attachment_content(attachments, "attachments")
-
-        assert file_name in result
-        assert result[file_name]["url"] == download_url
-        assert result[file_name]["size"] == expected_size
-
-    @pytest.mark.parametrize(
-        "file_name,download_url,file_size,expected_size",
-        [
-            ("test.txt", None, 1024, None),  # Missing URL should be invalid
-            (
-                "test.pdf",
-                "https://example.com/test.pdf",
-                None,
-                "Unknown",
-            ),  # Missing size is valid
-            (
-                None,
-                "https://example.com/test.txt",
-                1024,
-                None,
-            ),  # Missing filename should be invalid
-        ],
-    )
-    def test_write_attachment_content_no_content_parametrized(
-        self,
-        privacy_request: PrivacyRequest,
-        file_name,
-        download_url,
-        file_size,
-        expected_size,
-    ):
-        """Test attachment content writing with missing required fields"""
-        builder = DSRReportBuilder(privacy_request, {})
-        attachments = [
-            {
-                "file_name": file_name,
-                "download_url": download_url,
-                "file_size": file_size,
-            }
-        ]
-
-        result = builder._write_attachment_content(attachments, "attachments")
-
-        if expected_size is None:
-            # Should return empty dict for invalid attachments (missing filename or URL)
-            assert not result
-        else:
-            # Should include attachment with missing size
-            assert file_name in result
-            assert result[file_name]["url"] == download_url
-            assert result[file_name]["size"] == expected_size
-
-    @pytest.mark.parametrize("directory", ["attachments", "data/dataset1/collection1"])
-    def test_write_attachment_content_different_directories_parametrized(
-        self, privacy_request: PrivacyRequest, directory
-    ):
-        """Test attachment content writing in different directories"""
-        builder = DSRReportBuilder(privacy_request, {})
-        attachments = [
-            {
-                "file_name": "test.txt",
-                "download_url": "https://example.com/test.txt",
-                "file_size": 1024,
-            }
-        ]
-
-        result = builder._write_attachment_content(attachments, directory)
-
-        assert "test.txt" in result
-        assert result["test.txt"]["url"] == "https://example.com/test.txt"
-        assert result["test.txt"]["size"] == "1.0 KB"
 
 
 class TestDSRReportBuilderContent(TestDSRReportBuilderBase):
@@ -1467,7 +718,6 @@ class TestDSRReportBuilderOrganization(TestDSRReportBuilderBase):
 
     def test_performance_large_dataset(self, privacy_request: PrivacyRequest):
         """Test performance with a large dataset"""
-
         # Create a smaller dataset for performance testing
         # 100 collections with 100 items each instead of 1000x1000
         dsr_data = {
