+++ tests/ops/service/storage/streaming/test_smart_open_streaming_storage.py
@@ -1,5 +1,3 @@
-"""Tests for SmartOpenStreamingStorage class."""
-
 import json
 import time
 from io import BytesIO
@@ -32,6 +30,28 @@ class TestSmartOpenStreamingStorage:
         storage = SmartOpenStreamingStorage(mock_smart_open_client, chunk_size=4096)
         assert storage.chunk_size == 4096
 
+    def test_parse_storage_url(self, mock_smart_open_client):
+        """Test parsing storage URL."""
+        storage = SmartOpenStreamingStorage(mock_smart_open_client)
+        assert storage._parse_storage_url("s3://test-bucket/test-key") == (
+            "test-bucket",
+            "test-key",
+        )
+        assert storage._parse_storage_url(
+            "https://test-bucket.s3.amazonaws.com/test-key"
+        ) == ("test-bucket", "test-key")
+        assert storage._parse_storage_url(
+            "http://test-bucket.s3.amazonaws.com/test-key"
+        ) == ("test-bucket", "test-key")
+        assert storage._parse_storage_url("https://test-bucket.com/test-key") == (
+            "test-bucket.com",
+            "test-key",
+        )
+        assert storage._parse_storage_url("http://test-bucket.com/test-key") == (
+            "test-bucket.com",
+            "test-key",
+        )
+
     def test_convert_to_stream_zip_format(self, mock_smart_open_client):
         """Test conversion to stream_zip format."""
         storage = SmartOpenStreamingStorage(mock_smart_open_client)
@@ -50,7 +70,11 @@ class TestSmartOpenStreamingStorage:
         assert result[0][3] is not None  # _ZIP_32_TYPE instance
         assert list(result[0][4]) == [b'{"test": "data"}']
 
-    def test_collect_attachments(self, mock_smart_open_client):
+
+class TestSmartOpenStreamingStorageAttachments:
+    """Test SmartOpenStreamingStorage attachments related methods."""
+
+    def test_collect_attachments_and_validate(self, mock_smart_open_client):
         """Test collecting attachments from data."""
         storage = SmartOpenStreamingStorage(mock_smart_open_client)
 
@@ -74,7 +98,12 @@ class TestSmartOpenStreamingStorage:
             ],
         }
 
-        result = storage._collect_and_validate_attachments(data)
+        result = storage._collect_and_validate_attachments(
+            data,
+            used_filenames_data=set(),
+            used_filenames_attachments=set(),
+            processed_attachments={},
+        )
 
         assert len(result) == 2
         # Check that direct attachments are processed
@@ -100,7 +129,12 @@ class TestSmartOpenStreamingStorage:
             ]
         }
 
-        result = storage._collect_and_validate_attachments(data)
+        result = storage._collect_and_validate_attachments(
+            data,
+            used_filenames_data=set(),
+            used_filenames_attachments=set(),
+            processed_attachments={},
+        )
 
         assert len(result) == 2
         assert result[0].attachment.file_name == "doc1.pdf"
@@ -118,7 +152,12 @@ class TestSmartOpenStreamingStorage:
             ]
         }
 
-        result = storage._collect_and_validate_attachments(data)
+        result = storage._collect_and_validate_attachments(
+            data,
+            used_filenames_data=set(),
+            used_filenames_attachments=set(),
+            processed_attachments={},
+        )
 
         # Attachments without file_name are skipped by the processing logic
         assert len(result) == 0
@@ -141,7 +180,12 @@ class TestSmartOpenStreamingStorage:
             ]
         }
 
-        result = storage._collect_and_validate_attachments(data)
+        result = storage._collect_and_validate_attachments(
+            data,
+            used_filenames_data=set(),
+            used_filenames_attachments=set(),
+            processed_attachments={},
+        )
 
         assert len(result) == 1
         assert result[0].attachment.file_name == "doc1.pdf"
@@ -296,7 +340,12 @@ class TestSmartOpenStreamingStorage:
             ]
         }
 
-        result = storage._collect_and_validate_attachments(data)
+        result = storage._collect_and_validate_attachments(
+            data,
+            used_filenames_data=set(),
+            used_filenames_attachments=set(),
+            processed_attachments={},
+        )
 
         assert len(result) == 1
         assert result[0].attachment.file_name == "doc1.pdf"
