+++ tests/ops/service/storage/test_util.py
@@ -6,13 +6,10 @@ from pytest import param
 
 from fides.api.service.storage.util import (
     AllowedFileType,
-    convert_processed_attachments_to_attachment_processing_info,
-    extract_storage_key_from_attachment,
     get_allowed_file_type_or_raise,
     get_local_filename,
     get_unique_filename,
-    resolve_base_path_from_context,
-    resolve_directory_from_context,
+    resolve_path_from_context,
 )
 
 
@@ -247,113 +244,6 @@ class TestGetUniqueFilename:
         assert result == expected_result
 
 
-class TestExtractStorageKeyFromAttachment:
-    """Tests for the extract_storage_key_from_attachment function"""
-
-    @pytest.mark.parametrize(
-        "attachment, expected_key",
-        [
-            param(
-                {"original_download_url": "https://example.com/file.pdf"},
-                "https://example.com/file.pdf",
-                id="original_download_url_only",
-            ),
-            param(
-                {
-                    "original_download_url": "https://example.com/file.pdf",
-                    "download_url": "https://backup.com/file.pdf",
-                },
-                "https://example.com/file.pdf",
-                id="original_download_url_priority",
-            ),
-            param(
-                {"download_url": "https://example.com/file.pdf"},
-                "https://example.com/file.pdf",
-                id="download_url_only",
-            ),
-            param(
-                {
-                    "download_url": "https://example.com/file.pdf",
-                    "file_name": "backup.pdf",
-                },
-                "https://example.com/file.pdf",
-                id="download_url_priority_over_filename",
-            ),
-            param(
-                {"file_name": "document.pdf"},
-                "document.pdf",
-                id="file_name_only",
-            ),
-            param(
-                {
-                    "original_download_url": "",
-                    "download_url": "",
-                    "file_name": "document.pdf",
-                },
-                "document.pdf",
-                id="empty_urls_fallback_to_filename",
-            ),
-            param(
-                {
-                    "original_download_url": None,
-                    "download_url": None,
-                    "file_name": "document.pdf",
-                },
-                "document.pdf",
-                id="none_urls_fallback_to_filename",
-            ),
-            param(
-                {
-                    "original_download_url": "",
-                    "download_url": "",
-                    "file_name": "",
-                },
-                "",
-                id="all_empty_returns_empty",
-            ),
-            param(
-                {
-                    "original_download_url": None,
-                    "download_url": None,
-                    "file_name": None,
-                },
-                "",
-                id="all_none_returns_empty",
-            ),
-            param(
-                {
-                    "original_download_url": "s3://bucket/path/file.pdf",
-                    "download_url": "https://example.com/file.pdf",
-                    "file_name": "document.pdf",
-                },
-                "s3://bucket/path/file.pdf",
-                id="s3_url_priority",
-            ),
-        ],
-    )
-    def test_extract_storage_key(self, attachment, expected_key):
-        """Test that storage key is extracted correctly with proper fallback logic"""
-        result = extract_storage_key_from_attachment(attachment)
-        assert result == expected_key
-
-    def test_extract_storage_key_missing_keys(self):
-        """Test that missing keys are handled gracefully"""
-        result = extract_storage_key_from_attachment({})
-        assert result == ""
-
-    def test_extract_storage_key_with_extra_keys(self):
-        """Test that extra keys in attachment don't interfere"""
-        attachment = {
-            "original_download_url": "https://example.com/file.pdf",
-            "download_url": "https://backup.com/file.pdf",
-            "file_name": "document.pdf",
-            "extra_field": "should_be_ignored",
-            "size": 1024,
-        }
-        result = extract_storage_key_from_attachment(attachment)
-        assert result == "https://example.com/file.pdf"
-
-
 class TestGenerateAttachmentUrlFromStoragePath:
     """Tests for the generate_attachment_url_from_storage_path function"""
 
@@ -451,8 +341,8 @@ class TestGenerateAttachmentUrlFromStoragePath:
             assert result == expected_encoded, f"Failed for filename: {filename}"
 
 
-class TestResolveBasePathFromContext:
-    """Tests for the resolve_base_path_from_context function"""
+class TestResolvePathFromContext:
+    """Tests for the resolve_path_from_context function"""
 
     @pytest.mark.parametrize(
         "attachment, default_base_path, expected_path",
@@ -531,296 +421,13 @@ class TestResolveBasePathFromContext:
             ),
         ],
     )
-    def test_resolve_base_path(self, attachment, default_base_path, expected_path):
+    def test_resolve_path(self, attachment, default_base_path, expected_path):
         """Test that base path is resolved correctly based on context"""
-        result = resolve_base_path_from_context(attachment, default_base_path)
+        result = resolve_path_from_context(attachment, default_base_path)
         assert result == expected_path
 
-    def test_resolve_base_path_with_none_context(self):
+    def test_resolve_path_with_none_context(self):
         """Test that None context is handled gracefully"""
         attachment = {"_context": None}
-        result = resolve_base_path_from_context(attachment, "default")
+        result = resolve_path_from_context(attachment, "default")
         assert result == "default"
-
-
-class TestResolveDirectoryFromContext:
-    """Tests for the resolve_directory_from_context function"""
-
-    @pytest.mark.parametrize(
-        "attachment, default_directory, expected_directory",
-        [
-            param(
-                {
-                    "_context": {
-                        "type": "direct",
-                        "dataset": "users",
-                        "collection": "profiles",
-                    }
-                },
-                "attachments",
-                "data/users/profiles",
-                id="direct_context",
-            ),
-            param(
-                {
-                    "_context": {
-                        "type": "nested",
-                        "dataset": "orders",
-                        "collection": "items",
-                    }
-                },
-                "attachments",
-                "data/orders/items",
-                id="nested_context",
-            ),
-            param(
-                {"_context": {"type": "top_level"}},
-                "attachments",
-                "attachments",
-                id="top_level_context",
-            ),
-            param(
-                {
-                    "_context": {
-                        "type": "old_format",
-                        "key": "dataset:collection",
-                        "item_id": "123",
-                    }
-                },
-                "attachments",
-                "dataset:collection/123",
-                id="old_context_format",
-            ),
-            param(
-                {},  # no context
-                "attachments",
-                "attachments",
-                id="no_context_default",
-            ),
-            param(
-                {},  # no context with custom default
-                "custom/path",
-                "custom/path",
-                id="no_context_custom_default",
-            ),
-            param(
-                {"_context": {}},  # empty context
-                "attachments",
-                "attachments",
-                id="empty_context",
-            ),
-            param(
-                {
-                    "_context": {
-                        "type": "direct",
-                        "dataset": "users",
-                        "collection": "profiles",
-                    }
-                },
-                "custom/default",
-                "data/users/profiles",
-                id="context_overrides_default",
-            ),
-            param(
-                {
-                    "_context": {"key": "dataset:collection", "item_id": "123"}
-                },  # old format without type
-                "attachments",
-                "dataset:collection/123",
-                id="old_format_without_type",
-            ),
-            param(
-                {
-                    "_context": {"key": "dataset:collection"}
-                },  # old format missing item_id
-                "attachments",
-                "attachments",
-                id="old_format_missing_item_id",
-            ),
-        ],
-    )
-    def test_resolve_directory(self, attachment, default_directory, expected_directory):
-        """Test that directory is resolved correctly based on context"""
-        result = resolve_directory_from_context(attachment, default_directory)
-        assert result == expected_directory
-
-    def test_resolve_directory_with_none_context(self):
-        """Test that None context is handled gracefully"""
-        attachment = {"_context": None}
-        result = resolve_directory_from_context(attachment, "default")
-        assert result == "default"
-
-
-class TestConvertProcessedAttachmentsToAttachmentProcessingInfo:
-    """Tests for the convert_processed_attachments_to_attachment_processing_info function"""
-
-    def test_convert_with_valid_attachments(self):
-        """Test conversion with valid attachments"""
-
-        # Mock validation function that returns a simple object
-        def mock_validate_func(attachment_with_context):
-            return f"validated_{attachment_with_context['file_name']}"
-
-        processed_attachments_list = [
-            {
-                "attachment": {
-                    "file_name": "test1.pdf",
-                    "download_url": "https://example.com/1",
-                },
-                "context": {"type": "direct", "dataset": "users"},
-            },
-            {
-                "attachment": {
-                    "file_name": "test2.pdf",
-                    "download_url": "https://example.com/2",
-                },
-                "context": {"type": "nested", "dataset": "orders"},
-            },
-        ]
-
-        result = convert_processed_attachments_to_attachment_processing_info(
-            processed_attachments_list, mock_validate_func
-        )
-
-        assert len(result) == 2
-        assert result[0] == "validated_test1.pdf"
-        assert result[1] == "validated_test2.pdf"
-
-    def test_convert_with_none_validation_function(self):
-        """Test conversion with None validation function"""
-        processed_attachments_list = [
-            {
-                "attachment": {
-                    "file_name": "test1.pdf",
-                    "download_url": "https://example.com/1",
-                },
-                "context": {"type": "direct", "dataset": "users"},
-            },
-        ]
-
-        result = convert_processed_attachments_to_attachment_processing_info(
-            processed_attachments_list, None
-        )
-
-        assert len(result) == 0
-
-    def test_convert_with_validation_function_returning_none(self):
-        """Test conversion when validation function returns None for some attachments"""
-
-        def mock_validate_func(attachment_with_context):
-            # Only validate attachments that start with "valid"
-            if attachment_with_context["file_name"].startswith("valid"):
-                return f"validated_{attachment_with_context['file_name']}"
-            return None
-
-        processed_attachments_list = [
-            {
-                "attachment": {
-                    "file_name": "valid.pdf",
-                    "download_url": "https://example.com/1",
-                },
-                "context": {"type": "direct", "dataset": "users"},
-            },
-            {
-                "attachment": {
-                    "file_name": "invalid.pdf",
-                    "download_url": "https://example.com/2",
-                },
-                "context": {"type": "nested", "dataset": "orders"},
-            },
-            {
-                "attachment": {
-                    "file_name": "valid2.pdf",
-                    "download_url": "https://example.com/3",
-                },
-                "context": {"type": "top_level"},
-            },
-        ]
-
-        result = convert_processed_attachments_to_attachment_processing_info(
-            processed_attachments_list, mock_validate_func
-        )
-
-        assert len(result) == 2
-        assert result[0] == "validated_valid.pdf"
-        assert result[1] == "validated_valid2.pdf"
-
-    def test_convert_with_empty_list(self):
-        """Test conversion with empty processed attachments list"""
-
-        def mock_validate_func(attachment_with_context):
-            return f"validated_{attachment_with_context['file_name']}"
-
-        result = convert_processed_attachments_to_attachment_processing_info(
-            [], mock_validate_func
-        )
-
-        assert len(result) == 0
-
-    def test_convert_preserves_context_in_attachment(self):
-        """Test that context is properly added to attachment data"""
-        captured_attachments = []
-
-        def mock_validate_func(attachment_with_context):
-            captured_attachments.append(attachment_with_context)
-            return f"validated_{attachment_with_context['file_name']}"
-
-        processed_attachments_list = [
-            {
-                "attachment": {
-                    "file_name": "test.pdf",
-                    "download_url": "https://example.com/1",
-                },
-                "context": {
-                    "type": "direct",
-                    "dataset": "users",
-                    "collection": "profiles",
-                },
-            },
-        ]
-
-        convert_processed_attachments_to_attachment_processing_info(
-            processed_attachments_list, mock_validate_func
-        )
-
-        assert len(captured_attachments) == 1
-        captured = captured_attachments[0]
-        assert captured["file_name"] == "test.pdf"
-        assert captured["download_url"] == "https://example.com/1"
-        assert captured["_context"] == {
-            "type": "direct",
-            "dataset": "users",
-            "collection": "profiles",
-        }
-
-    def test_convert_does_not_modify_original_attachment(self):
-        """Test that original attachment data is not modified"""
-        original_attachment = {
-            "file_name": "test.pdf",
-            "download_url": "https://example.com/1",
-        }
-        original_context = {"type": "direct", "dataset": "users"}
-
-        def mock_validate_func(attachment_with_context):
-            # Modify the attachment to test it doesn't affect the original
-            attachment_with_context["modified"] = True
-            return f"validated_{attachment_with_context['file_name']}"
-
-        processed_attachments_list = [
-            {
-                "attachment": original_attachment,
-                "context": original_context,
-            },
-        ]
-
-        convert_processed_attachments_to_attachment_processing_info(
-            processed_attachments_list, mock_validate_func
-        )
-
-        # Verify original data is unchanged
-        assert original_attachment == {
-            "file_name": "test.pdf",
-            "download_url": "https://example.com/1",
-        }
-        assert original_context == {"type": "direct", "dataset": "users"}
-        assert "modified" not in original_attachment
